feat(application): Updated landing page and application system for cl… (#54)
* feat(application): Updated landing page and application system for closed portal

Refactored the application to handle the scenario when the portal is closed. Updated the landing page to display a new message when the application portal is closed. Also, added a new attribute 'closedPortal' in multiple places to check and display vital information accordingly. A new function 'getLastActiveHackathonId' is introduced to get the last active hackathon when the portal is closed.

* fix? Update src/server/getters/landingPageInfo.ts

Co-authored-by: Matej Tarča <matejtarca@gmail.com>

* fix: Update src/scenes/Application/components/ApplicationForm/ApplicationFormSteps.tsx

Co-authored-by: Matej Tarča <matejtarca@gmail.com>

* fix: Refactor handling of closed portal in application

Restructured code for improved handling of events when the portal is closed. Introduced a fallback to get the last active hackathon ID for closed portals, improving reliability in displaying hackathon-specific information.

* fix: Simplify handling of 'closedPortal' state

The variable 'closedPortal' has been removed to directly set the 'closedPortal' state to the boolean value of 'activeHackathonId'. This simplifies the condition checks and improves code readability.

* fix: Refactor hackathon session retrieval and update hackathon sorting

Refactor the `getHackerForActiveHackathonFromSession` function to improve code readability with more comments and simpler conditional logic. Update the `getLastActiveHackathonId` function to sort hackathons directly in the database query rather than in JavaScript, providing performance benefits.

* fix: Refactor sorting method in getLastActiveHackathonId

Updated the "getLastActiveHackathonId" function to sort hackathons by event start date in descending order directly in the database query. This update enhances performance by eliminating the need for sorting in JavaScript and removes unnecessarily imported Prisma SortOrder.

* Refactor hackathon retrieval in getHackerForActiveHackathonFromSession

This commit revises the method for retrieving hackathons in the getHackerForActiveHackathonFromSession function. The 'activeHackathonId' and 'lastActiveHackathonId' variables have been consolidated into a single 'hackathonId' variable. This streamlines the process of checking for active hackathons and closing the portal if none are found.

---------

Co-authored-by: Matej Tarča <matejtarca@gmail.com>